+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added `post-install` manifest option support.
 - Rename `rebuild` command to `build-and-run`. It would also now do a build automatically without having to run a separate `build` command.
 - Don't require for the build to be initialized when running `clean` command.
+- Add Export bundle command
 
 ## [0.0.24]